• DataReaderDialogFactory now uses a string map instead of a DataReader pointer since that object was only used in the map • remove setXMLFileName from modules • Various warning removals

+++ src/qt/editor/linkdialog.cpp
@@ -725,7 +725,7 @@ void LinkDialogGraphicsScene::addPropertyLink(Property* sProp, Property* eProp, 
 }
 
 int LinkDialogGraphicsScene::currentLinkItemsCount() {
-    return currentConnectionGraphicsItems_.size();
+    return static_cast<int>(currentConnectionGraphicsItems_.size());
 }
 
 void LinkDialogGraphicsScene::addPropertyLink(PropertyLink* propertyLink) {
